github-control/requirements.txt
flask==3.0.0
gitpython==3.1.40
python-dotenv==1.0.0
flask-bootstrap==3.3.7.1
